refactor(backend): Refactor boss module

+++ server/src/boss/services/boss-progress-service.js
@@ -1,19 +1,23 @@
 import BossProgressModel from "../models/boss-progress-model.js";
 import { validateObjectId } from "../helpers/boss-helpers.js";
 
+const createDefaultProgress = async (userId) => {
+    return BossProgressModel.create({
+        user: userId,
+        lastDefeatedBossId: 0,
+        currentAvailableBossId: 1,
+        totalBossesDefeated: 0,
+        totalExpFromBosses: 0,
+    });
+};
+
 const getBossProgress = async (userId) => {
     validateObjectId(userId, "user ID");
 
     let progress = await BossProgressModel.findOne({ user: userId });
 
     if (!progress) {
-        progress = await BossProgressModel.create({
-            user: userId,
-            lastDefeatedBossId: 0,
-            currentAvailableBossId: 1,
-            totalBossesDefeated: 0,
-            totalExpFromBosses: 0,
-        });
+        progress = await createDefaultProgress(userId);
     }
 
     return progress;
@@ -24,10 +28,12 @@ const updateBossProgress = async (userId, defeatedBossId, expGained) => {
 
     const progress = await getBossProgress(userId);
 
-    progress.lastDefeatedBossId = defeatedBossId;
-    progress.currentAvailableBossId = defeatedBossId + 1;
-    progress.totalBossesDefeated += 1;
-    progress.totalExpFromBosses += expGained;
+    Object.assign(progress, {
+        lastDefeatedBossId: defeatedBossId,
+        currentAvailableBossId: defeatedBossId + 1,
+        totalBossesDefeated: progress.totalBossesDefeated + 1,
+        totalExpFromBosses: progress.totalExpFromBosses + expGained,
+    });
     await progress.save();
 
     return progress;
@@ -38,9 +44,7 @@ const resetBossProgress = async (userId) => {
 
     const progress = await BossProgressModel.findOne({ user: userId });
 
-    if (progress) {
-        await progress.deleteOne();
-    }
+    if (progress) await progress.deleteOne();
 
     return true;
 };
